+++ Entity.h
@@ -25,7 +25,7 @@ enum KEYS : int {none = 0, UP = 1, LEFT = 2, DOWN = 3, RIGHT = 4};
         ~Entity() {};
     
     public:
-        Entity(int x = 0, int y = 0, int vx = 0, int vy = 0, int wh = 0, int hh = 0, ALLEGRO_BITMAP* p = nullptr):
+        Entity(float x = 0, float y = 0, float vx = 0, float vy = 0, int wh = 0, int hh = 0, ALLEGRO_BITMAP* p = nullptr):
             cord_x(x), cord_y(y), velocity_x(vx), velocity_y(vy), width_hitbox(wh), height_hitbox(hh), image(p) {}
 
         void print(ALLEGRO_BITMAP* buffer) const; //da implementare
