+++ Player.h
@@ -11,7 +11,7 @@
 
     protected:
     public:
-        Player(unsigned l=1, AXIS s=X, KEYS p=none, int x=0, int y=0, int vx=0, int vy=0, int wh=0, int hh=0, ALLEGRO_BITMAP* p=nullptr)
+        Player(unsigned l=1, AXIS s=X, KEYS p=none, float x=0, float y=0, float vx=0, float vy=0, int wh=0, int hh=0, ALLEGRO_BITMAP* p=nullptr)
             : Entity(x, y, vx, vy, wh, hh, p), lifes(l), is_safe(s), position(p) {}
         
         inline bool move(KEYS k);
